new file

+++ ABC/123C.cpp
@@ -6,17 +6,17 @@ using namespace std;
 
 int main()
 {
-    int N; cin >> N;
-    vector <int> vec(5);
+    unsigned long N; cin >> N;
+    vector <unsigned long> vec(5);
 
     for (int i = 0; i < 5; i++){
         cin >> vec.at(i);
     }
 
     sort(vec.begin(),vec.end());
-    int output = vec.at(0);
+    unsigned long output = vec.at(0);
 
-    int t = N/output;
+    unsigned long t = N/output;
 
     if (N % output == 0){
         cout << 4 + t << endl;
